// src/app/app-enum.ts
export enum Url {
    LOGIN = 'login',
    INSCRIRE = 'inscrire',
    STATISTIQUES = 'statistiques',
    ARTICLES = 'articles',
    ARTICLE = 'article',
    MVTSTK = 'mvtstk',
    CLIENT = 'client',
    FOURNISSEUR = 'fournisseur',
    CLIENTS = 'clients',
    FOURNISSEURS = 'fournisseurs',
    COMMANDES_CLIENTS = 'commandes-clients',
    COMMANDES_FOURNISSEURS = 'commandes-fournisseurs',
    COMMANDE_CLIENT = 'commande-client',
    COMMANDE_FOURNISSEUR = 'commande-fournisseur',
    CATEGORIES = 'categories',
    CATEGORIE = 'categorie',
    UTILISATEURS = 'utilisateurs',
    UTILISATEUR = 'utilisateur',
    PROFIL = 'profil'
}